Move tasks to chamber_of_deputies
